# docs/swagger.yaml
definitions:
  main.Category:
    properties:
      description:
        example: Deskripsi Kategori 1
        type: string
      id:
        type: integer
      name:
        example: Kategori 1
        type: string
    type: object
  main.Produk:
    properties:
      harga:
        example: 2000
        type: integer
      id:
        example: 1
        type: integer
      nama:
        example: Kopi Gadjah
        type: string
      stok:
        example: 10
        type: integer
    type: object
host: kasir-api-production-1064.up.railway.app
info:
  contact: {}
  description: Ini adalah API untuk sistem kasir sederhana.
  title: Kasir API
  version: "1.0"
paths:
  /api/categories:
    get:
      description: Get list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Category'
            type: array
        "404":
          description: Category belum ada
          schema:
            type: string
      summary: Get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category with the provided data
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/main.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Add a new category
      tags:
      - Category
  /api/categories/{id}:
    delete:
      description: Remove a category from the list by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: category deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Category ID
          schema:
            type: string
        "404":
          description: Category tidak ditemukan
          schema:
            type: string
      summary: Delete a category
      tags:
      - Category
    get:
      description: Get details of a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Invalid Category ID
          schema:
            type: string
        "404":
          description: Produk belum ada
          schema:
            type: string
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update category details by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/main.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Category'
        "400":
          description: Invalid request/ID
          schema:
            type: string
        "404":
          description: Produk tidak ditemukan
          schema:
            type: string
      summary: Update an existing category
      tags:
      - Category
  /api/produk:
    get:
      description: Get list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Produk'
            type: array
        "404":
          description: Produk belum ada
          schema:
            type: string
      summary: Get all products
      tags:
      - Produk
    post:
      consumes:
      - application/json
      description: Create a new product with the provided data
      parameters:
      - description: Product data
        in: body
        name: produk
        required: true
        schema:
          $ref: '#/definitions/main.Produk'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Produk'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Add a new product
      tags:
      - Produk
  /api/produk/{id}:
    delete:
      description: Remove a product from the list by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: produk deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid produk ID
          schema:
            type: string
        "404":
          description: Produk tidak ditemukan
          schema:
            type: string
      summary: Delete a product
      tags:
      - Produk
    get:
      description: Get details of a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Produk'
        "400":
          description: Invalid produk ID
          schema:
            type: string
        "404":
          description: Produk belum ada
          schema:
            type: string
      summary: Get product by ID
      tags:
      - Produk
    put:
      consumes:
      - application/json
      description: Update product details by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: produk
        required: true
        schema:
          $ref: '#/definitions/main.Produk'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Produk'
        "400":
          description: Invalid request/ID
          schema:
            type: string
        "404":
          description: Produk tidak ditemukan
          schema:
            type: string
      summary: Update an existing product
      tags:
      - Produk
  /health:
    get:
      description: Get the status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check API Health
      tags:
      - General
schemes:
- https
- http
swagger: "2.0"